pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "tagorganizer"
license = {file = "COPYING"}
dynamic = ["version"]
description = "Organize photos and videos using hierachical tags"
readme = "README.md"
authors = [
  { name="Arun Persaud", email="arun@nubati.net" },
]
maintainers = [
  { name="Arun Persaud", email="arun@nubati.net" },
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "docopt>=0.6.2",
    "exifread>=3.0.0",
    "folium>=0.19.3",
    "matplotlib>=3.10.0",
    "more-itertools>=10.5.0",
    "numpy>=2.2.1",
    "opencv-python-headless>=4.11.0.86",
    "platformdirs>=4.3.6",
    "pyqt6>=6.8.0",
    "pyqt6-webengine>=6.8.0",
    "python-dateutil>=2.9.0.post0",
    "python-vlc>=3.0.21203",
    "qtpy>=2.4.2",
    "sqlmodel>=0.0.22",
    "xxhash>=3.5.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Topic :: Multimedia :: Graphics :: Viewers",
]
[project.urls]
Homepage = "https://github.com/arunpersaud/TagOrganizer"
Repository = "https://github.com/arunpersaud/TagOrganizer"
Issues = "https://github.com/arunpersaud/TagOrganizer/issues"
Changelog = "https://github.com/arunpersaud/TagOrganizer/blob/main/CHANGELOG.md"

[project.scripts]
TagOrganizer = "tagorganizer.main:main"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tagorganizer = ["alembic.ini", "migrations/script.py.mako"]
